Adding the condition to start a game
