+++ pkg/data/queue.go
@@ -4,6 +4,8 @@ import (
 	"fmt"
 )
 
+const numPlayersInAGame = 4
+
 // ErrorPlayerNotFound : Player specific errors
 var ErrorPlayerNotFound = fmt.Errorf("Player not found")
 
@@ -19,12 +21,13 @@ type Players []*Player
 
 // InQueue returns a boolean that verifies that a player is queued
 func InQueue(id int) bool {
-	return findIndexByPlayerID(id) > 0
+	return findIndexByPlayerID(id) > -1
 }
 
 // AddPlayer append a player to the queue
-func AddPlayer(player *Player) {
+func AddPlayer(player *Player) error {
 	queue = append(queue, player)
+	return checkQueue()
 }
 
 // DeletePlayer deletes the player with the given id from the queue
@@ -48,6 +51,27 @@ func findIndexByPlayerID(id int) int {
 		}
 	}
 	return -1
+}
+
+// Check if the number of players in the queue is sufficient to start a game
+// If so, a request with a group of players is sent to the Dispatcher to start a game
+// Complexity will increase in the next iteration
+func checkQueue() error {
+	fmt.Println(queue)
+	if len(queue) >= numPlayersInAGame {
+		party := queue[0:numPlayersInAGame-1]
+		err := startGame(party)
+		if err != nil {
+			return err
+		}
+		queue = queue[numPlayersInAGame:]
+	}
+	return nil
+}
+
+func startGame(party Players) error {
+	// Send request with party to Dispatcher
+	return nil
 }
 
 var queue = []*Player{
