+++ pkg/data/validate_test.go
@@ -34,38 +34,11 @@ func TestPlayerAlreadyInQueue(t *testing.T) {
 		UserIP: "123.123.123.123",
 	}
 
+	AddPlayer(player)
+
 	err := player.ValidatePlayer()
 
 	if !(err != nil && err.Error() == "Key: 'Player.UserID' Error:Field validation for 'UserID' failed on the 'notinqueue' tag") {
 		t.Fatal("The player is assumed to be already in the queue and should not be added again")
-	}
-}
-
-func TestAddPlayer(t *testing.T) {
-	player := &Player{
-		UserID: 1,
-		UserIP: "123.123.123.123",
-	}
-
-	AddPlayer(player)
-
-	if findIndexByPlayerID(player.UserID) == -1 {
-		t.Fatal("The player has not been added to the queue")
-	}
-}
-
-func TestDeletePlayer(t *testing.T) {
-	err := DeletePlayer(42)
-
-	if err != nil {
-		t.Fatal(err)
-	}
-}
-
-func TestDeletePlayerNotInQueue(t *testing.T) {
-	err := DeletePlayer(2)
-
-	if !(err != nil && err.Error() == "Player not found") {
-		t.Fatal("The player is not in the queue, so it is assumed not to be found")
 	}
 }
